refactor(schema): remove AggregatedConditionDefinitionProvider
- Remove AbstractAggregatedQueryDefinitionProvider and related classes
- Update AggregatedCondition interface and implementation
- Adjust AggregatedConditionCapableDefinitionProvider to use new condition structure
- Remove unused OrderAggregatedCondition.json schema
- Update related test resources and configurations
