--- wow-schema/src/main/kotlin/me/ahoo/wow/schema/typed/query/AbstractAggregatedQueryDefinitionProvider.kt
@@ -1,46 +1,0 @@
-/*
- * Copyright [2021-present] [ahoo wang <ahoowang@qq.com> (https://github.com/Ahoo-Wang)].
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *      http://www.apache.org/licenses/LICENSE-2.0
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package me.ahoo.wow.schema.typed.query
-
-import com.fasterxml.classmate.ResolvedType
-import com.github.victools.jsonschema.generator.CustomDefinition
-import com.github.victools.jsonschema.generator.SchemaGenerationContext
-import me.ahoo.wow.schema.JsonSchema
-import me.ahoo.wow.schema.JsonSchema.Companion.asCustomDefinition
-import me.ahoo.wow.schema.JsonSchema.Companion.asJsonSchema
-
-abstract class AbstractAggregatedQueryDefinitionProvider : QueryDefinitionProvider {
-    override fun provideCustomSchemaDefinition(
-        javaType: ResolvedType,
-        context: SchemaGenerationContext
-    ): CustomDefinition? {
-        if (!javaType.isInstanceOf(aggregatedType)) {
-            return null
-        }
-        val schemaVersion = context.generatorConfig.schemaVersion
-        val queryType = context.typeContext.resolve(queryType)
-        val rootSchema = context.createStandardDefinition(queryType, this).asJsonSchema(schemaVersion)
-        if (javaType.typeBindings.isEmpty) {
-            return rootSchema.asCustomDefinition()
-        }
-        val commandAggregateType = javaType.typeBindings.getBoundType(0).erasedType
-        return createCustomDefinition(rootSchema, commandAggregateType, context)
-    }
-
-    abstract fun createCustomDefinition(
-        rootSchema: JsonSchema,
-        commandAggregateType: Class<*>,
-        context: SchemaGenerationContext
-    ): CustomDefinition
-}
