+++ wow-schema/src/test/resources/META-INF/wow-schema/OrderAggregatedCondition.json
@@ -1,46 +1,12 @@
 {
   "$schema" : "https://json-schema.org/draft/2020-12/schema",
-  "$defs" : {
-    "StringObjectMap" : {
-      "type" : "object"
-    },
-    "wow.api.query.Condition" : {
-      "type" : "object",
-      "properties" : {
-        "children" : {
-          "default" : "[]",
-          "type" : "array",
-          "items" : {
-            "$ref" : "#/$defs/wow.api.query.Condition"
-          }
-        },
-        "field" : {
-          "type" : "string"
-        },
-        "operator" : {
-          "$ref" : "#/$defs/wow.api.query.Operator",
-          "default" : "ALL"
-        },
-        "options" : {
-          "$ref" : "#/$defs/StringObjectMap",
-          "default" : "{}"
-        },
-        "value" : { }
-      },
-      "required" : [ "operator" ]
-    },
-    "wow.api.query.Operator" : {
-      "type" : "string",
-      "enum" : [ "AND", "OR", "NOR", "ID", "IDS", "AGGREGATE_ID", "AGGREGATE_IDS", "TENANT_ID", "OWNER_ID", "DELETED", "ALL", "EQ", "NE", "GT", "LT", "GTE", "LTE", "CONTAINS", "IN", "NOT_IN", "BETWEEN", "ALL_IN", "STARTS_WITH", "ENDS_WITH", "ELEM_MATCH", "NULL", "NOT_NULL", "TRUE", "FALSE", "EXISTS", "TODAY", "BEFORE_TODAY", "TOMORROW", "THIS_WEEK", "NEXT_WEEK", "LAST_WEEK", "THIS_MONTH", "LAST_MONTH", "RECENT_DAYS", "EARLIER_DAYS", "RAW" ]
-    }
-  },
   "type" : "object",
   "properties" : {
     "children" : {
       "default" : "[]",
       "type" : "array",
       "items" : {
-        "$ref" : "#/$defs/wow.api.query.Condition"
+        "$ref" : "#"
       }
     },
     "field" : {
@@ -48,14 +14,15 @@
       "enum" : [ "", "aggregateId", "tenantId", "ownerId", "version", "eventId", "firstOperator", "operator", "firstEventTime", "eventTime", "deleted", "state", "state.address", "state.address.city", "state.address.country", "state.address.detail", "state.address.district", "state.address.province", "state.id", "state.items", "state.items.id", "state.items.price", "state.items.productId", "state.items.quantity", "state.items.totalPrice", "state.paidAmount", "state.payable", "state.status", "state.totalAmount" ]
     },
     "operator" : {
-      "$ref" : "#/$defs/wow.api.query.Operator",
+      "type" : "string",
+      "enum" : [ "AND", "OR", "NOR", "ID", "IDS", "AGGREGATE_ID", "AGGREGATE_IDS", "TENANT_ID", "OWNER_ID", "DELETED", "ALL", "EQ", "NE", "GT", "LT", "GTE", "LTE", "CONTAINS", "IN", "NOT_IN", "BETWEEN", "ALL_IN", "STARTS_WITH", "ENDS_WITH", "ELEM_MATCH", "NULL", "NOT_NULL", "TRUE", "FALSE", "EXISTS", "TODAY", "BEFORE_TODAY", "TOMORROW", "THIS_WEEK", "NEXT_WEEK", "LAST_WEEK", "THIS_MONTH", "LAST_MONTH", "RECENT_DAYS", "EARLIER_DAYS", "RAW" ],
       "default" : "ALL"
     },
     "options" : {
-      "$ref" : "#/$defs/StringObjectMap",
+      "type" : "object",
       "default" : "{}"
     },
     "value" : { }
   },
-  "required" : [ "operator" ]
+  "required" : [ "children", "field", "operator", "value" ]
 }
