+++ wow-schema/src/test/resources/META-INF/wow-schema/OrderAggregatedListQuery.json
@@ -4,6 +4,32 @@
     "StringObjectMap" : {
       "type" : "object"
     },
+    "example.OrderAggregatedCondition" : {
+      "type" : "object",
+      "properties" : {
+        "children" : {
+          "default" : "[]",
+          "type" : "array",
+          "items" : {
+            "$ref" : "#/$defs/example.OrderAggregatedCondition"
+          }
+        },
+        "field" : {
+          "type" : "string",
+          "enum" : [ "", "aggregateId", "tenantId", "ownerId", "version", "eventId", "firstOperator", "operator", "firstEventTime", "eventTime", "deleted", "state", "state.address", "state.address.city", "state.address.country", "state.address.detail", "state.address.district", "state.address.province", "state.id", "state.items", "state.items.id", "state.items.price", "state.items.productId", "state.items.quantity", "state.items.totalPrice", "state.paidAmount", "state.payable", "state.status", "state.totalAmount" ]
+        },
+        "operator" : {
+          "$ref" : "#/$defs/wow.api.query.Operator",
+          "default" : "ALL"
+        },
+        "options" : {
+          "$ref" : "#/$defs/StringObjectMap",
+          "default" : "{}"
+        },
+        "value" : { }
+      },
+      "required" : [ "children", "field", "operator", "value" ]
+    },
     "wow.api.query.Condition" : {
       "type" : "object",
       "properties" : {
@@ -37,30 +63,7 @@
   "type" : "object",
   "properties" : {
     "condition" : {
-      "type" : "object",
-      "properties" : {
-        "children" : {
-          "default" : "[]",
-          "type" : "array",
-          "items" : {
-            "$ref" : "#/$defs/wow.api.query.Condition"
-          }
-        },
-        "field" : {
-          "type" : "string",
-          "enum" : [ "", "aggregateId", "tenantId", "ownerId", "version", "eventId", "firstOperator", "operator", "firstEventTime", "eventTime", "deleted", "state", "state.address", "state.address.city", "state.address.country", "state.address.detail", "state.address.district", "state.address.province", "state.id", "state.items", "state.items.id", "state.items.price", "state.items.productId", "state.items.quantity", "state.items.totalPrice", "state.paidAmount", "state.payable", "state.status", "state.totalAmount" ]
-        },
-        "operator" : {
-          "$ref" : "#/$defs/wow.api.query.Operator",
-          "default" : "ALL"
-        },
-        "options" : {
-          "$ref" : "#/$defs/StringObjectMap",
-          "default" : "{}"
-        },
-        "value" : { }
-      },
-      "required" : [ "operator" ]
+      "$ref" : "#/$defs/example.OrderAggregatedCondition"
     },
     "limit" : {
       "type" : "integer",
